+++ .github/CODEOWNERS
@@ -45,7 +45,6 @@ crates/bitwarden-crypto-sync-handler/** @bitwarden/team-key-management-dev
 crates/bitwarden-exporters/** @bitwarden/team-tools-dev
 crates/bitwarden-generators/** @bitwarden/team-tools-dev
 crates/bitwarden-importers/** @bitwarden/team-tools-dev
-crates/bitwarden-km-sync-handler/** @bitwarden/team-key-management-dev
 crates/bitwarden-organization-crypto/** @bitwarden/team-key-management-dev
 crates/bitwarden-organization-invite-link/** @bitwarden/team-admin-console-dev
 crates/bitwarden-organizations/** @bitwarden/team-admin-console-dev
