--- crates/bitwarden-km-sync-handler/uniffi.toml
@@ -1,11 +1,0 @@
-[bindings.kotlin]
-package_name = "com.bitwarden.kmsynchandler"
-generate_immutable_records = true
-android = true
-# Temporary workaround for https://github.com/mozilla/uniffi-rs/issues/2740
-omit_checksums = true
-
-[bindings.swift]
-ffi_module_name = "BitwardenKmSyncHandlerFFI"
-module_name = "BitwardenKmSyncHandler"
-generate_immutable_records = true
