+++ crates/bitwarden-uniffi/Cargo.toml
@@ -34,7 +34,6 @@ bitwarden-exporters = { workspace = true, features = ["uniffi"] }
 bitwarden-fido = { workspace = true, features = ["uniffi"] }
 bitwarden-generators = { workspace = true, features = ["uniffi"] }
 bitwarden-importers = { workspace = true, features = ["uniffi"] }
-bitwarden-km-sync-handler = { workspace = true, features = ["uniffi"] }
 bitwarden-organizations = { workspace = true, features = ["uniffi"] }
 bitwarden-pm = { workspace = true, features = ["uniffi"] }
 bitwarden-policies = { workspace = true, features = ["uniffi"] }
